requirements/local.txt
# Import all base requirements
-r base.txt

# Django debug toolbar
django-debug-toolbar==1.6
